+++ view/html-metabox-pinned-tasks.php
@@ -18,7 +18,6 @@ require_once $ptc_completionist->plugin_path . 'src/class-html-builder.php';
 
 try {
 
-  Asana_Interface::require_license();
   Asana_Interface::require_settings();
   $asana = Asana_Interface::get_client();